Refactor custom view logic out of ui and into repl.
