+++ src/sidekick/exceptions.py
@@ -1,0 +1,4 @@
+class SidekickAbort(Exception):
+    """Exception raised to abort the current sidekick operation."""
+
+    pass
