+++ src/sidekick/tools/run_command.py
@@ -1,6 +1,5 @@
 import subprocess
 
-
 message_handler = None
