+++ src/sidekick/ui.py
@@ -1,5 +1,3 @@
-import json
-
 from prompt_toolkit.formatted_text import HTML
 from rich.console import Console
 from rich.markdown import Markdown
@@ -9,7 +7,7 @@ from rich.pretty import Pretty
 from rich.table import Table
 
 from sidekick import config, session
-from sidekick.utils.helpers import DotDict, ext_to_lang, key_to_title, render_file_diff
+from sidekick.utils.helpers import DotDict
 
 BANNER = """\
 ███████╗██╗██████╗ ███████╗██╗  ██╗██╗ ██████╗██╗  ██╗
@@ -81,6 +79,10 @@ def error(text: str):
     panel("Error", text, style=colors.error)
 
 
+def markdown(text: str):
+    return Markdown(text)
+
+
 def dump_messages():
     messages = Pretty(session.messages)
     panel("Message History", messages, style=colors.muted)
@@ -135,141 +137,3 @@ def show_help():
         table.add_row(cmd, desc)
 
     panel("Available Commands", table, border_style=colors.muted)
-
-
-def _create_code_block(filepath: str, content: str) -> Markdown:
-    """
-    Create a code block for the given file path and content.
-
-    Args:
-        filepath (str): The path to the file.
-        content (str): The content of the file.
-
-    Returns:
-        Markdown: A Markdown object representing the code block.
-    """
-    lang = ext_to_lang(filepath)
-    code_block = f"```{lang}\n{content}\n```"
-    return Markdown(code_block)
-
-
-def _render_args(tool_name, args):
-    """
-    Render the tool arguments for a given tool.
-
-    """
-    # Show diff between `target` and `patch` on file updates
-    if tool_name == "update_file":
-        return render_file_diff(args["target"], args["patch"], colors)
-
-    # Show file content on read_file
-    elif tool_name == "write_file":
-        return _create_code_block(args["filepath"], args["content"])
-
-    # Default to showing key and value on new line
-    content = ""
-    for key, value in args.items():
-        if isinstance(value, list):
-            content += f"{key_to_title(key)}:\n"
-            for item in value:
-                content += f"  - {item}\n"
-            content += "\n"
-        else:
-            # If string length is over 200 characters
-            # split to new line
-            # content += f"{key.title()}:\n{value}\n\n"
-            value = str(value)
-            content += f"{key_to_title(key)}:"
-            if len(value) > 200:
-                content += f"\n{value}\n\n"
-            else:
-                content += f" {value}\n\n"
-    return content.strip()
-
-
-def _parse_args(args):
-    """
-    Parse tool arguments from a JSON string or dictionary.
-
-    Args:
-        args (str or dict): A JSON-formatted string or a dictionary containing tool arguments.
-
-    Returns:
-        dict: The parsed arguments.
-
-    Raises:
-        ValueError: If 'args' is not a string or dictionary, or if the string is not valid JSON.
-    """
-    if isinstance(args, str):
-        try:
-            return json.loads(args)
-        except json.JSONDecodeError:
-            raise ValueError(f"Invalid JSON: {args}")
-    elif isinstance(args, dict):
-        return args
-    else:
-        raise ValueError(f"Invalid args type: {type(args)}")
-
-
-def _log_mcp(title, args):
-    """Display MCP tool with its arguments."""
-    if not args:
-        return
-
-    status(title)
-    for key, value in args.items():
-        if isinstance(value, list):
-            value = ", ".join(value)
-        muted(f"{key}: {value}", spaces=4)
-
-
-def _get_tool_title(tool_name):
-    """
-    Checks if the tool exists within this system. If it does
-    it return "Tool" otherwise assumed to be an MCP so returns "MCP"
-    """
-    if tool_name in config.INTERNAL_TOOLS:
-        return f"Tool({tool_name})"
-    else:
-        return f"MCP({tool_name})"
-
-
-def confirm(tool_call, node):
-    title = _get_tool_title(tool_call.tool_name)
-    args = _parse_args(tool_call.args)
-
-    # If we're skipping confirmation on this tool, log its output if MCP
-    if (
-        session.yolo
-        or tool_call.tool_name in session.tool_ignore
-        # or tool_call.tool_name in session.user_config["settings"]["tool_ignore"]
-    ):
-        if tool_call.tool_name not in config.INTERNAL_TOOLS:
-            _log_mcp(title, args)
-        return
-
-    session.spinner.stop()
-    content = _render_args(tool_call.tool_name, args)
-    filepath = args.get("filepath")
-
-    # Set bottom padding to 0 if filepath is not None
-    bottom_padding = 0 if filepath else 1
-
-    panel(title, content, bottom=bottom_padding, border_style=colors.warning)
-
-    # If tool call has filepath, show it under panel
-    if filepath:
-        show_usage(f"File: {filepath}")
-
-    print("  1. Yes (default)")
-    print("  2. Yes, and don't ask again for commands like this")
-    print("  3. No, and tell Sidekick what to do differently")
-    resp = input("  Choose an option [1/2/3]: ").strip() or "1"
-
-    if resp == "2":
-        session.tool_ignore.append(tool_call.tool_name)
-    elif resp == "3":
-        raise UserAbort("User aborted.")
-
-    line()  # Add line after user input
-    session.spinner.start()
